UPSTREAM: drm/i915/dp_mst: Make HBLANK expansion quirk work for logical ports
The DPCD OUI of the logical port on a Dell UHBR monitor - on which the
AUX device is used to enable DSC - is all 0. To detect if the HBLANK
expansion quirk is required for this monitor use the OUI of the port's
parent instead.

Since in the above case the DPCD of both the logical port and the parent
port reports being a sink device (vs. branch device) type, read the
proper sink/branch OUI based on the DPCD device type.

This is required by a follow-up patch enabling the quirk for the above
Dell monitor.

Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240416221010.376865-11-imre.deak@intel.com
(cherry picked from commit 427c70302bbef0542a11464066447d66f97a6fe0)

BUG=b:332903004, b:322223603
TEST=Verify Dell U3224KB 6k mode on Rex using UHBR, MST encoder, Big
     Joiner, and DSC

Change-Id: Id278780fd98e1fe079a9bbbbe3458dfea4dcdd6c
Signed-off-by: Rakshith M O <rakshith.m.o@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5802891
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Manasi Navare <navaremanasi@chromium.org>