CHROMIUM: iwl7000: ship changes/code only for relevant kernels
Even in recent kernels, we have a lot of changes that are
simply not needed, and make it appear that the driver is
vastly different from upstream and our internal versions.

Rejigger the tooling that builds the ChromeOS version of
our driver, so that the driver is no longer identical on
different Chrome kernel versions, but rather tailored to
the specific kernel version.

This means that e.g. on older kernels we now completely
remove some code that cannot work there (rather than just
adding ifdefs around it), and on newer kernels we have a
lot less unnecessary changes. For example, where struct
members are accessed that don't exist on all versions, we
have accessors that would then depend on the version, but
it's nicer to have the code without that indirection, so
on kernels that have it, this is now left along, and on
kernels that don't, it's just replaced directly. That bit
sometimes causes strange code to appear, but that seems a
reasonable trade-off.

Change-Id: I51b408107fdde8e294552e038ed617a800641d97
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b77b3abe38fbca6f74859072306a147ca5e8d691