FROMGIT: r8152: Wake up the system if the we need a reset
If we get to the end of the r8152's suspend() routine and we find that
the USB device is INACCESSIBLE then it means that some of our
preparation for suspend didn't take place. We need a USB reset to get
ourselves back in a consistent state so we can try again and that
can't happen during system suspend. Call pm_wakeup_event() to wake the
system up in this case.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Hayes Wang <hayeswang@realtek.com>
Link: https://lore.kernel.org/r/66590f25.170a0220.8b5ad.1752@mx.google.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 8c1d92a740c002f36a919a55689d0ef83b61fd1c
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/ main)

BUG=b:340925255
TEST=Fake failure in suspend; r8152 recovers

Change-Id: I8b91ce1c783ba11c6c3722396035c737e5b7b12b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5596395
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Daniel Matei <danielgeorgem@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>