CHROMIUM: sched/eevdf: Add ENFORCE_ELIGIBILITY
Eligibility reduces scheduling latency of high nice task at the
cost of runtime of low nice tasks. This hurts the chrome model where
UI tasks should not be interrupted by background tasks. This adds
a feature to be able to disable eligibility checks.

Note that fairness is presvered in the way the deadline is calculated
since the deadline is adjusted by the weight of the task and a task.

BUG=b:308209790
TEST=Manual tested on DUT
UPSTREAM-TASK=b:324602237

Change-Id: Ie69ad1b2cb08d91d20cbc74076a8c8949a7ac823
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5670701
Tested-by: Youssef Esmat <youssefesmat@chromium.org>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>