CHROMIUM: arm64: dts: mediatek: mt8186: adjust MALI GPU mapping
This just changes the order of the clocks specified so that we can use
the same dts with _either_ the downstream MALI driver or the upstream
panfrost driver. The downstream MALI driver uses the clock-names so
the order doesn't matter. The upstream panfrost driver just expects
one clock and it seems to be "&mfgsys CLK_MFG_BG3D".

This should probably be squashed into the commit ("CHROMIUM: arm64:
mt8186: dts: add proprietary clock names") if/when rebasing.

BUG=b:351768680
UPSTREAM-TASK=b:260376782
TEST=Corsola graphics w/ downstream MALI driver still comes up

Change-Id: I6dae75b0b89d3067e358111ebc9beebcb175f1d0
Signed-off-by: Douglas Anderson <dianders@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5811744
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Wojciech Macek <wmacek@google.com>
Commit-Queue: Albert Jakieła <jakiela@google.com>
Reviewed-by: Wojciech Macek <wmacek@google.com>
Reviewed-by: Albert Jakieła <jakiela@google.com>