BACKPORT: FROMLIST: drm/msm: fix the highest_bank_bit for sc7180
sc7180 programs the ubwc settings as 0x1e as that would mean a
highest bank bit of 14 which matches what the GPU sets as well.

However, the highest_bank_bit field of the msm_mdss_data which is
being used to program the SSPP's fetch configuration is programmed
to a highest bank bit of 16 as 0x3 translates to 16 and not 14.

Fix the highest bank bit field used for the SSPP to match the mdss
and gpu settings.

Fixes: 6f410b246209 ("drm/msm/mdss: populate missing data")
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Reviewed-by: Rob Clark <robdclark@gmail.com>
Tested-by: Stephen Boyd <swboyd@chromium.org> # Trogdor.Lazor
(am from https://patchwork.kernel.org/patch/13758255/)
(also found at https://lore.kernel.org/r/20240808235227.2701479-1-quic_abhinavk@quicinc.com)

Conflicts:
   drivers/gpu/drm/msm/msm_mdss.c

We don't have commit a55c8ff252d3 ("drm/msm/mdss: Handle the reg bus ICC
path"). This will come via stable kernel but we're landing it FROMLIST
to see if it fixes $BUG in the field.

BUG=b:332414449
UPSTREAM-TASK=b:359266252
TEST=DP on Lazor

Change-Id: Id54460d299062ddeb697ebc2d5d89946ec2fa217
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5782228
Reviewed-by: Rob Clark <robdclark@chromium.org>