Revert "UPSTREAM: drm/i915: Rename the DSM/GSM registers"
This reverts commit 4ea9f45a01dfe164646663c58346463780ac33df as
android is not booting with MTL in GVT-d mode.

When android is launched using qemu with trusty enabled, guest check
will return as false due to signature not matching with KVM cpuid
signature. This will result in stolen memory access via the BARs
which will result in kernel panic.

Reverted the changes to fix MTL boot issue in GVT-d mode.

Tests done:
- Android boot in GVT-d and BM config
- adb reboot

Tracked-On: OAM-123278
Change-Id: I2cefa412b38e581f6d45670318cb2b7d26e12666
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>