Revert "UPSTREAM: drm/i915: Disable the "binder""
This reverts commit 4098b4fd915c606d3b898ec8bc8bc603e9cf5816 as
android is not booting with MTL in GVT-d mode.

When android is launched using qemu with trusty enabled, guest check
will return as false due to signature not matching with KVM cpuid
signature. This will result in stolen memory access via the BARs
which will result in kernel panic.

Reverted the changes to fix MTL boot issue in GVT-d mode.

Tests done:
- Android boot in GVT-d and BM config
- adb reboot

Tracked-On: OAM-123278
Change-Id: I5a45a578b9569c05a03daae77da6f02390c7670b
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>