Revert "UPSTREAM: drm/i915: Bypass LMEMBAR/GTTMMADR for MTL stolen memory access"
This reverts commit bc4185cf460aa2aca4ce63ef672ee4defafe58f0 as
android is not booting with MTL in GVT-d mode.

When android is launched using qemu with trusty enabled, guest check
will return as false due to signature not matching with KVM cpuid
signature. This will result in stolen memory access via the BARs
which will result in kernel panic.

Reverted the changes to fix MTL boot issue in GVT-d mode.

Tests done:
- Android boot in GVT-d and BM config
- adb reboot

Tracked-On: OAM-123278
Change-Id: I0d63b984c8875a5bea6359594d22d961f56c470a
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>