CHROMIUM: HID: multitouch: Scope mt_reset actions to haptic devices
To handle haptic device reset, mt_reset was added to PM functions.
In kernel 6.6, the calls to mt_reset seems to negatively impact an
increasing number of devices across different manufacturers. To avoid
continuously adding devices to the quirk: SKIP_RESET, we are limiting
the reset actions to only haptic devices. This is a temporary workaround
whiles we come up with a long term robust fix that will be part
of the effort to upstream support for haptic touchpads.

BUG=b:352329398
TEST=Build and test kernel on affected devices
UPSTREAM-TASK=b:298170707

Change-Id: I74bc44408b7b44a199e0a50b60fe444a9da71329
Signed-off-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5895129
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Angela Czubak <aczubak@google.com>